[Pal] Refactor and rename global structures
The new names should better reflect what these structures really are.
This commit also makes passing of the initial state to the start ecall
more explicit (previously `pal_sec` was partially used for this and
partially for storing other global state).

The next step will be moving topology information to a separate struct
and getting rid of `pal_sec`.

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>